+++ src/lib/st-form/st-form.module.ts
@@ -23,9 +23,10 @@ import { StFormFieldComponent } from './st-form-field/st-form-field.component';
 import { PipesModule } from '../pipes/pipes.module';
 import { StInputModule } from '../st-input/st-input.module';
 import { StFormDirectiveModule } from '../directives/form/form-directives.module';
+import { StSwitchModule } from '../st-switch/st-switch.module';
 
 @NgModule({
-   imports: [CommonModule, FormsModule, ReactiveFormsModule, StInputModule, PipesModule, StFormDirectiveModule],
+   imports: [CommonModule, FormsModule, ReactiveFormsModule, StInputModule, StSwitchModule, PipesModule, StFormDirectiveModule],
    declarations: [StFormComponent, StFormFieldComponent],
    exports: [StFormComponent]
 })