+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * st-breadcrumbs: Add component st-breadcrumbs item, add Ng-Content option
 * st-input: Add input value and output change
 * st-form: Create a dynamic form with inputs
+* st-form: Add switches to dynamic form
 
 **Refactor:**